Add Guid field to Note, add public empty ctor
Until this commit, the Guid conecpt is only used locally, meaning a Note
only has a Guid on a local machine (determined by its filename).
However, considering synchronisation, a Note should be identified by a
Guid beyond machine boundaries. I.e., in a syncing concept two Notes,
which are identically in terms of having the same Guid, can exist with
different content (because their sync revisions are different). Thus we
have to store and transmit the Guid everytime we access a note.